Extract debugconsole to own module
