+++ src/robotide/__init__.py
@@ -30,8 +30,6 @@ release. The most stable, and best documented, module is `robotide.pluginapi`.
 import sys
 import os
 from string import Template
-import traceback
-import threading
 
 errorMessageTemplate = Template("""$reason
 You need to install wxPython $versions toolkit with unicode support to run RIDE.
@@ -41,30 +39,36 @@ supported_versions = ["2.8"]
 try:
     import wxversion
     from wxversion import VersionError
-    if sys.platform == 'darwin': # CAN NOT IMPORT IS_MAC AS THERE IS A wx IMPORT
+    if sys.platform == 'darwin':
         supported_versions.append("2.9")
     wxversion.select(supported_versions)
     import wx
     if "ansi" in wx.PlatformInfo:
-        print errorMessageTemplate.substitute(reason="wxPython with ansi encoding is not supported", versions=" or ".join(supported_versions))
+        print errorMessageTemplate.substitute(
+            reason="wxPython with ansi encoding is not supported",
+            versions=" or ".join(supported_versions))
         sys.exit(1)
 except ImportError as e:
-    if "no appropriate 64-bit architecture" in e.message.lower() and sys.platform == 'darwin':
-        print "python should be executed in 32-bit mode to support wxPython on mac. Check BUILD.rest for details"
+    if "no appropriate 64-bit architecture" in e.message.lower() and \
+       sys.platform == 'darwin':
+        print "python should be executed in 32-bit mode with wxPython on OSX."
     else:
-        print errorMessageTemplate.substitute(reason="wxPython not found.", versions=" or ".join(supported_versions))
+        print errorMessageTemplate.substitute(
+            reason="wxPython not found.",
+            versions=" or ".join(supported_versions))
     sys.exit(1)
 except VersionError:
-    print errorMessageTemplate.substitute(reason="Wrong wxPython version.", versions=" or ".join(supported_versions))
+    print errorMessageTemplate.substitute(
+        reason="Wrong wxPython version.",
+        versions=" or ".join(supported_versions))
     sys.exit(1)
 
 # Insert bundled robot to path before anything else
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'lib'))
+sys.path.append(os.path.join(os.path.dirname(__file__), 'spec'))
 
 from robotide.robotapi import DataError
-
-
-sys.path.append(os.path.join(os.path.dirname(__file__), 'spec'))
+from robotide.application import debugconsole
 
 
 def main(*args):
@@ -77,13 +81,16 @@ def main(*args):
     except DataError, err:
         print str(err) + '\n\nUse --help to get usage information.'
 
+
 def _parse_args(args):
     if not args:
         return False, False, None
     noupdatecheck = '--noupdatecheck' in args
     debug_console = '--debugconsole' in args
-    inpath = args[-1] if args[-1] not in ['--noupdatecheck', '--debugconsole'] else None
+    inpath = args[-1] if args[-1] not in ['--noupdatecheck', '--debugconsole'] \
+        else None
     return noupdatecheck, debug_console, inpath
+
 
 def _run(inpath=None, updatecheck=True, debug_console=False):
     try:
@@ -96,8 +103,9 @@ def _run(inpath=None, updatecheck=True, debug_console=False):
     ride = RIDE(inpath, updatecheck)
     _show_old_wxpython_warning_if_needed(ride.frame)
     if debug_console:
-        _start_debug_console(ride)
+        debugconsole.start(ride)
     ride.MainLoop()
+
 
 def _show_old_wxpython_warning_if_needed(parent=None):
     if wx.VERSION >= (2, 8, 12, 1):
@@ -112,25 +120,10 @@ def _show_old_wxpython_warning_if_needed(parent=None):
                % wx.VERSION_STRING)
     style = wx.ICON_EXCLAMATION
     if not parent:
-        app = wx.PySimpleApp()
+        _ = wx.PySimpleApp()
         parent = wx.Frame(None, size=(0,0))
     wx.MessageDialog(parent, message, title, style).ShowModal()
 
-def _print_stacks():
-    id2name = dict((th.ident, th.name) for th in threading.enumerate())
-    for threadId, stack in sys._current_frames().items():
-        print(id2name[threadId])
-        traceback.print_stack(f=stack)
-
-def _start_debug_console(ride):
-    import code
-    help_string = """\
-RIDE - access to the running application
-print_stacks() - print current stack traces
-"""
-    console = code.InteractiveConsole(locals={'RIDE':ride, 'print_stacks':_print_stacks})
-    thread = threading.Thread(target=lambda: console.interact(help_string))
-    thread.start()
 
 if __name__ == '__main__':
     main(sys.argv[1:])
